+++ spec/support/form_helpers.rb
@@ -293,7 +293,7 @@ module FormHelpers
     choose("activity[covid19_related]", option: covid19_related)
     click_button t("form.button.activity.submit")
 
-    if parent&.is_gcrf_funded?
+    if parent&.is_gcrf_funded? || parent&.roda_identifier_fragment == "GCRF"
       expect(page).to have_content t("form.legend.activity.gcrf_challenge_area")
       expect(page).to have_content t("form.hint.activity.gcrf_challenge_area")
       choose("activity[gcrf_challenge_area]", option: gcrf_challenge_area)