+++ terraform/variables.tf
@@ -95,3 +95,13 @@ variable "rollbar_disabled" {
   description = "Flag to turn off rollbar reporting"
   default     = "false"
 }
+
+variable "custom_domain" {
+  type        = string
+  description = "Name of custom domain created in the cf org"
+}
+
+variable "custom_hostname" {
+  type        = string
+  description = "Custom hostname (prepended to custom_domain for the app and cdn-route)"
+}